# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Bounty API Documentation",
    "version": "1.0.0",
    "description": "API documentation for the Bounty application",
    "contact": {
      "name": "VS",
      "email": "vishal@33-sol.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "http://localhost:5001/",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token in the format **Bearer &lt;token>**"
      }
    },
    "schemas": {
      "Campaign": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "60f7c0b5d4d3c81234567890"
          },
          "user": {
            "type": "string",
            "example": "60f7c0b5d4d3c81234567890"
          },
          "company": {
            "$ref": "#/components/schemas/Company"
          },
          "name": {
            "type": "string",
            "example": "Festive Campaign"
          },
          "description": {
            "type": "string",
            "example": "Campaign for the festive season with QR code rewards."
          },
          "totalAmount": {
            "type": "number",
            "example": 1000
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "festive",
              "holiday"
            ]
          },
          "status": {
            "type": "string",
            "example": "Pending"
          },
          "messageTemplate": {
            "type": "string",
            "example": "Welcome to our campaign!"
          },
          "zipUrl": {
            "type": "string",
            "example": "http://example.com/path/to/zipfile.zip"
          },
          "payoutConfig": {
            "type": "object",
            "description": "Payout configuration settings"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Company": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "60f7c0b5d4d3c81234567890"
          },
          "name": {
            "type": "string",
            "example": "ABC Corp"
          },
          "user": {
            "type": "string",
            "example": "60f7c0b5d4d3c81234567890"
          },
          "plan": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "example": "Elite Pro"
              },
              "qrCodeLimit": {
                "type": "integer",
                "example": 1000
              },
              "qrCodesGenerated": {
                "type": "integer",
                "example": 100
              }
            }
          },
          "whatsappNumber": {
            "type": "string",
            "example": "+1234567890"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/campaigns": {
      "get": {
        "summary": "Get all campaigns for the authenticated user",
        "tags": [
          "Campaigns"
        ],
        "responses": {
          "200": {
            "description": "A list of campaigns",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "campaigns": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Campaign"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/api/companies": {
      "post": {
        "summary": "Create a new company",
        "tags": [
          "Company"
        ],
        "description": "Allows the creation of a company in the system.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "contactName": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "phoneNumber": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "website": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Company created successfully"
          },
          "400": {
            "description": "Error in company creation data"
          }
        }
      }
    },
...
    "/auth/google": {
      "post": {
        "summary": "Authenticate user via Google",
        "description": "Authenticates a user using their Google ID token and returns a JWT for session management.",
        "tags": [
          "GOOGLE SSO"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "idToken": {
                    "type": "string",
                    "description": "Google ID token obtained from the client-side after Google Sign-In.",
                    "example": "eyJhbGciOiJSUzI1NiIsImtpZCI6Ij..."
                  }
                },
                "required": [
                  "idToken"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful. Returns JWT and user information.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "description": "JSON Web Token (JWT) for authenticated sessions.",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "description": "Unique identifier for the user.",
                          "example": "60d5ec49f1d2c916c8f0e8b7"
                        },
                        "full_name": {
                          "type": "string",
                          "description": "Full name of the user.",
                          "example": "John Doe"
                        },
                        "email": {
                          "type": "string",
                          "description": "Email address of the user.",
                          "example": "john.doe@example.com"
                        },
                        "avatar": {
                          "type": "string",
                          "description": "URL to the user's avatar image.",
                          "example": "https://example.com/avatar/johndoe.jpg"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request. Missing or invalid ID token.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Invalid or missing ID token."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. Failed to authenticate the ID token.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Invalid ID token."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error. An unexpected error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "An unexpected error occurred."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/login": {
      "post": {
        "summary": "User login",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "phoneNumber": {
                    "type": "string",
                    "example": "918779780352"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
  },
...
}